refactor: rename StatementPdfGenerator to StatementPdfWriter
